Fixed bugs connected with rename dialog (close/cancel dialog without changing name of directory, ability of creation file with extension). Fixed bug connected with deleting already deleted directory/file and updating panel that contain it as current directory

+++ FileSystem.cs
@@ -146,9 +146,9 @@ namespace FileManager.Entities
         public static void Delete(List<Item> items)
         {
             foreach (Item x in items)
-                if (x.File != null)
+                if (x.File != null && File.Exists(x.File.FullName))
                     x.File.Delete();
-                else
+                else if (Directory.Exists(x.Directory.FullName))
                     x.Directory.Delete(true);
         }
